ISSUE-62: Moved SportType to new namespace
