+++ src/Domain/Strava/Activity/Activity.php
@@ -6,10 +6,10 @@ use App\Domain\Measurement\Length\Kilometer;
 use App\Domain\Measurement\Length\Meter;
 use App\Domain\Measurement\Velocity\KmPerHour;
 use App\Domain\Nominatim\Location;
+use App\Domain\Strava\Activity\SportType\SportType;
 use App\Domain\Strava\Activity\Stream\PowerOutput;
 use App\Domain\Strava\Gear\GearId;
 use App\Domain\Strava\LeafletMap;
-use App\Domain\Strava\SportType;
 use App\Domain\Weather\OpenMeteo\Weather;
 use App\Infrastructure\Eventing\RecordsEvents;
 use App\Infrastructure\Time\TimeFormatter;
