+++ src/Domain/Strava/Activity/SportType/SportTypesToImport.php
@@ -2,9 +2,8 @@
 
 declare(strict_types=1);
 
-namespace App\Domain\Strava\Activity;
+namespace App\Domain\Strava\Activity\SportType;
 
-use App\Domain\Strava\SportType;
 use App\Infrastructure\ValueObject\Collection;
 
 final class SportTypesToImport extends Collection
